llm_completo/requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0

# Data
pandas>=2.0.0
pyarrow>=14.0.0

# Azure
azure-storage-file-datalake>=12.14.0
azure-identity>=1.15.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0

# LLM
anthropic>=0.18.0
groq>=0.4.0
scikit-learn>=1.4.0

# MCP
mcp>=1.0.0

# Utils
rich>=13.0.0
pydantic>=2.5.0
